requirements.txt
dotmap==1.3.26
numpy==1.21.2
pandas==1.4.1
Pillow==9.0.1
PyYAML==6.0
requests==2.27.1
scikit-learn==1.0.2
scipy==1.8.0
setuptools==58.0.4
sklearn==0.0
torch==1.11.0
torchvision==0.12.0
tornado==6.1
tqdm==4.63.1
wandb==0.12.11
